works better when I actually add the files

+++ main.c
@@ -28,15 +28,15 @@ int main(int argc, char* argv[]){
    // positions and speed at time n-1 (0) and n (1).
    vector p10,p20,v10,v20,p11,p21;
    // masses
-   double m1,m2;
+   long double m1,m2;
    // times
-   double dt, t_final, dt_output;
+   long double dt, t_final, dt_output;
 
-   fscanf(inputFile, "%lf\r\n%lf %lf %lf\r\n%lf %lf %lf\r\n%lf\r\n%lf %lf %lf\r\n%lf %lf %lf\r\n%lf\r\n%lf\r\n%lf", &m1, &(p10.x), &(p10.y), &(p10.z), &(v10.x), &(v10.y), &(v10.z), &m2, &(p20.x), &(p20.y), &(p20.z), &(v20.x), &(v20.y), &(v20.z), &dt, &t_final, &dt_output );
+   fscanf(inputFile, "%Lf\r\n%Lf %Lf %Lf\r\n%Lf %Lf %Lf\r\n%Lf\r\n%Lf %Lf %Lf\r\n%Lf %Lf %Lf\r\n%Lf\r\n%Lf\r\n%Lf", &m1, &(p10.x), &(p10.y), &(p10.z), &(v10.x), &(v10.y), &(v10.z), &m2, &(p20.x), &(p20.y), &(p20.z), &(v20.x), &(v20.y), &(v20.z), &dt, &t_final, &dt_output );
    
    if(verbatim){
       printf("Read parameters\r\n-------------\r\n");
-      printf("m1 \t %lf\r\n p1 X \t %lf\r\n p1 Y \t %lf\r\n p1 Z  \t %lf\r\n p1 U  \t %lf\r\n p1 V  \t %lf\r\n p1 W \t %lf\r\n m2 \t %lf\r\n p2 X \t %lf\r\n p2 Y \t %lf\r\n p2 Z \t %lf\r\n p2 U \t %lf\r\n p2 V \t %lf\r\n p2 W \t %lf\r\n dt \t %lf\r\n endT \t %lf\r\n outT \t %lf\r\n", m1, p10.x, p10.y, p10.z, v10.x, v10.y, v10.z, m2, p20.x, p20.y, p20.z, v20.x, v20.y, v20.z, dt,t_final,dt_output);
+      printf("m1 \t %Lf\r\n p1 X \t %Lf\r\n p1 Y \t %Lf\r\n p1 Z  \t %Lf\r\n p1 U  \t %Lf\r\n p1 V  \t %Lf\r\n p1 W \t %Lf\r\n m2 \t %Lf\r\n p2 X \t %Lf\r\n p2 Y \t %Lf\r\n p2 Z \t %Lf\r\n p2 U \t %Lf\r\n p2 V \t %Lf\r\n p2 W \t %Lf\r\n dt \t %Lf\r\n endT \t %Lf\r\n outT \t %Lf\r\n", m1, p10.x, p10.y, p10.z, v10.x, v10.y, v10.z, m2, p20.x, p20.y, p20.z, v20.x, v20.y, v20.z, dt,t_final,dt_output);
    }
 
    // Initial values determination
@@ -49,10 +49,10 @@ int main(int argc, char* argv[]){
       printf("\r\nComputed Data\r\n-------------\r\n");
    }
    
-   double next_out = dt_output;
-   for(double t = dt * 2 ; t < t_final ; t+= dt ){
+   long double next_out = dt_output;
+   for(long double t = dt * 2 ; t < t_final ; t+= dt ){
       vector dp = subv(p21,p11);
-      double norm = normv(dp);
+      long double norm = normv(dp);
       
       // compute F
       vector f = scalev(dp, - m1 * m2 * CONST_G / norm * norm * norm); 
@@ -67,8 +67,8 @@ int main(int argc, char* argv[]){
       p21 = new_p2;
       
       if(next_out <= t + 0.00001){
-         fprintf(outputFile, "%lf\t%lf\t%lf\t%lf\t%lf\t%lf\t%lf\r\n", t, p11.x, p11.y, p11.z, p21.x,p21.y, p21.z);
-         if(verbatim) printf("%lf\t%lf\t%lf\t%lf\t%lf\t%lf\t%lf\r\n", t, p11.x, p11.y, p11.z, p21.x,p21.y, p21.z);
+         fprintf(outputFile, "%Lf\t%Lf\t%Lf\t%Lf\t%Lf\t%Lf\t%Lf\r\n", t, p11.x, p11.y, p11.z, p21.x,p21.y, p21.z);
+         if(verbatim) printf("%Lf\t%Lf\t%Lf\t%Lf\t%Lf\t%Lf\t%Lf\r\n", t, p11.x, p11.y, p11.z, p21.x,p21.y, p21.z);
          next_out +=dt_output;
       }
    }
