+++ main_Matlab.c
@@ -43,11 +43,11 @@ int main(int argc, char* argv[]){
    // positions and speed at time n-1 (0) and n (1).
    vector p10,p20,v1,v2,p11,p21;
    // masses
-   double m1,m2;
+   long double m1,m2;
    // times
-   double dt, t_final, dt_output;
+   long double dt, t_final, dt_output;
 
-   fscanf(input_file, "%lf\r\n%lf %lf %lf\r\n%lf %lf %lf\r\n%lf\r\n%lf %lf %lf\r\n%lf %lf %lf\r\n%lf\r\n%lf\r\n%lf", &m1, &(p10.x), &(p10.y), &(p10.z), &(v1.x), &(v1.y), &(v1.z), &m2, &(p20.x), &(p20.y), &(p20.z), &(v2.x), &(v2.y), &(v2.z), &dt, &t_final, &dt_output );
+   fscanf(input_file, "%Lf\r\n%Lf %Lf %Lf\r\n%Lf %Lf %Lf\r\n%Lf\r\n%Lf %Lf %Lf\r\n%Lf %Lf %Lf\r\n%Lf\r\n%Lf\r\n%Lf", &m1, &(p10.x), &(p10.y), &(p10.z), &(v1.x), &(v1.y), &(v1.z), &m2, &(p20.x), &(p20.y), &(p20.z), &(v2.x), &(v2.y), &(v2.z), &dt, &t_final, &dt_output );
 
    // Initial values determination
    p11 = addv(scalev(v1, dt), p10);
@@ -58,11 +58,11 @@ int main(int argc, char* argv[]){
    write_output(output_file, matlab_file, use_verbatim, 0, p10, p20);
    if(dt_output <= dt + DOUBLE_EPSILON) write_output(output_file, matlab_file, use_verbatim, dt, p11, p21);
    
-   double next_out = dt_output;
+   long double next_out = dt_output;
  
-   for(double t = dt * 2 ; t <= t_final ; t+= dt ){
+   for(long double t = dt * 2 ; t <= t_final ; t+= dt ){
       vector dp = subv(p21,p11);
-      double norm = normv(dp);
+      long double norm = normv(dp);
       
       // compute F
       vector f = scalev(dp, - m1 * m2 * CONST_G / norm * norm * norm); 
@@ -70,7 +70,6 @@ int main(int argc, char* argv[]){
       // compute positions at time t
       vector new_p1 = addv(subv(scalev(f, - dt * dt / m1), p10) , scalev(p11, 2));
       vector new_p2 = addv(subv(scalev(f, dt * dt / m2), p20) , scalev(p21, 2));
-      
       p10 = p11;
       p20 = p21;
       p11 = new_p1;
@@ -89,16 +88,16 @@ int main(int argc, char* argv[]){
    return EXIT_SUCCESS;
 }
 
-void write_init(FILE* out1, FILE* out2, boolean verb, double m1, vector p1, vector v1, double m2, vector p2, vector v2, double dt, double t_final, double dt_output){
+void write_init(FILE* out1, FILE* out2, boolean verb, long double m1, vector p1, vector v1, long double m2, vector p2, vector v2, long double dt, long double t_final, long double dt_output){
    if(verb){
       printf("Read parameters\r\n-------------\r\n");
-      printf("m1 \t %lf\r\n p1 X \t %lf\r\n p1 Y \t %lf\r\n p1 Z  \t %lf\r\n p1 U  \t %lf\r\n p1 V  \t %lf\r\n p1 W \t %lf\r\n m2 \t %lf\r\n p2 X \t %lf\r\n p2 Y \t %lf\r\n p2 Z \t %lf\r\n p2 U \t %lf\r\n p2 V \t %lf\r\n p2 W \t %lf\r\n dt \t %lf\r\n endT \t %lf\r\n outT \t %lf\r\n", m1, p1.x, p1.y, p1.z, v1.x, v1.y, v1.z, m2, p2.x, p2.y, p2.z, v2.x, v2.y, v2.z, dt,t_final,dt_output);
+      printf("m1 \t %Lf\r\n p1 X \t %Lf\r\n p1 Y \t %Lf\r\n p1 Z  \t %Lf\r\n p1 U  \t %Lf\r\n p1 V  \t %Lf\r\n p1 W \t %Lf\r\n m2 \t %Lf\r\n p2 X \t %Lf\r\n p2 Y \t %Lf\r\n p2 Z \t %Lf\r\n p2 U \t %Lf\r\n p2 V \t %Lf\r\n p2 W \t %Lf\r\n dt \t %Lf\r\n endT \t %Lf\r\n outT \t %Lf\r\n", m1, p1.x, p1.y, p1.z, v1.x, v1.y, v1.z, m2, p2.x, p2.y, p2.z, v2.x, v2.y, v2.z, dt,t_final,dt_output);
       printf("\r\nInitial conditions\r\n------------\r\n");
-      printf(" p1 X(1) \t %f \r\n p1 Y(1) \t %f \r\n p1 Z(1) \t %f \r\n p2 X(1) \t %f \r\n p2 Y(1) \t %f \r\n p2 Z(1) \t %f \r\n", p1.x, p1.y, p1.z, p2.x, p2.y, p2.z);
+      printf(" p1 X(1) \t %Lf \r\n p1 Y(1) \t %Lf \r\n p1 Z(1) \t %Lf \r\n p2 X(1) \t %Lf \r\n p2 Y(1) \t %Lf \r\n p2 Z(1) \t %Lf \r\n", p1.x, p1.y, p1.z, p2.x, p2.y, p2.z);
       printf("\r\nComputed Data\r\n-------------\r\n");
    }
    if(out2 != NULL){
-      fprintf(out2, "% Matlab simulation script\r\n\r\nt_end = %lf;\r\ndt_computation = %lf;\r\ndt_sample = %lf;\r\n\r\nm1 = %lf; m2 = %lf;\r\n\r\nsamples = [\r\n", t_final, dt, dt_output, m1, m2);
+      fprintf(out2, "% Matlab simulation script\r\n\r\nt_end = %Lf;\r\ndt_computation = %Lf;\r\ndt_sample = %Lf;\r\n\r\nm1 = %Lf; m2 = %Lf;\r\n\r\nsamples = [\r\n", t_final, dt, dt_output, m1, m2);
    }
 }
 
@@ -110,8 +109,8 @@ void write_end(FILE* out1, FILE* out2, boolean verb){
    }
 }
 
-void write_output(FILE* out1, FILE* out2, boolean verb, double t, vector p1, vector p2){
-   fprintf(out1, "%lf\t%lf\t%lf\t%lf\t%lf\t%lf\t%lf\r\n", t, p1.x, p1.y, p1.z, p2.x,p2.y, p2.z);
-   if(verb) printf("%lf\t%lf\t%lf\t%lf\t%lf\t%lf\t%lf\r\n", t, p1.x, p1.y, p1.z, p2.x,p2.y, p2.z);
-   if(out2 != NULL) fprintf(out2, "%lf,%lf,%lf,%lf,%lf,%lf,%lf;\r\n", t, p1.x, p1.y, p1.z, p2.x,p2.y, p2.z);
+void write_output(FILE* out1, FILE* out2, boolean verb, long double t, vector p1, vector p2){
+   fprintf(out1, "%Lf\t%Lf\t%Lf\t%Lf\t%Lf\t%Lf\t%Lf\r\n", t, p1.x, p1.y, p1.z, p2.x,p2.y, p2.z);
+   if(verb) printf("%Lf\t%Lf\t%Lf\t%Lf\t%Lf\t%Lf\t%Lf\r\n", t, p1.x, p1.y, p1.z, p2.x,p2.y, p2.z);
+   if(out2 != NULL) fprintf(out2, "%Lf,%Lf,%Lf,%Lf,%Lf,%Lf,%Lf;\r\n", t, p1.x, p1.y, p1.z, p2.x,p2.y, p2.z);
 }
